apps,du: widen int to int64_t in parse_q_hyst_db and parse_q_offset_range
Same narrowing-wrap concern that motivated parse_bounded_int's int64_t
widening: with val.get<int>(), a JSON value larger than INT_MAX
(or smaller than INT_MIN) silently static_casts onto a 32-bit int and
can land on a valid switch case label, getting accepted as a different
valid value (e.g. JSON 4294967296 -> int 0 -> q_hyst_t::db0).

Read into int64_t so the original value survives long enough for the
switch to reject it via the default case, then cast to the target enum
once we know it matches a valid label.

No call-site changes. static_cast<q_hyst_t>(int64_t) and
static_cast<q_offset_range_t>(int64_t) are well-defined for the values
that actually reach the cast (each value in the validated case set).

+++ apps/units/flexible_o_du/o_du_high/du_high/commands/du_high_remote_commands.cpp
@@ -261,7 +261,7 @@ static expected<q_hyst_t, std::string> parse_q_hyst_db(const nlohmann::json& obj
   if (!key->is_number_integer()) {
     return make_unexpected("'q_hyst_db' value type should be an integer");
   }
-  const int v = key->get<int>();
+  const int64_t v = key->get<int64_t>();
   switch (v) {
     case 0:
     case 1:
@@ -296,7 +296,7 @@ static expected<q_offset_range_t, std::string> parse_q_offset_range(const nlohma
   if (!val.is_number_integer()) {
     return make_unexpected(fmt::format("'{}' value type should be an integer", field_name));
   }
-  const int v = val.get<int>();
+  const int64_t v = val.get<int64_t>();
   switch (v) {
     case -24:
     case -22:
